+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 This project uses [Gitmoji](https://gitmoji.carloscuesta.me/)
 
+## 1.0.1
+
+-   :wrench: update ESLint config adding React rules.
+-   :wrench: update `tsconfig.json` with React in mind.
+
 ## 1.0.0
 
 -   :sparkles: Initial commit.